DATAES-407 - improved tests for rest/transport template
