+++ src/test/java/org/springframework/data/elasticsearch/core/ElasticsearchTransportTemplateTests.java
@@ -1,0 +1,29 @@
+package org.springframework.data.elasticsearch.core;
+
+import static org.apache.commons.lang.RandomStringUtils.*;
+
+import org.elasticsearch.action.index.IndexRequest;
+import org.elasticsearch.common.xcontent.XContentType;
+import org.elasticsearch.index.engine.DocumentMissingException;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.data.elasticsearch.core.query.UpdateQuery;
+import org.springframework.data.elasticsearch.core.query.UpdateQueryBuilder;
+import org.springframework.data.elasticsearch.entities.SampleEntity;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@ContextConfiguration("classpath:elasticsearch-template-test.xml")
+public class ElasticsearchTransportTemplateTests extends ElasticsearchTemplateTests {
+
+	@Test(expected = DocumentMissingException.class)
+	public void shouldThrowExceptionIfDocumentDoesNotExistWhileDoingPartialUpdate() {
+		// when
+		IndexRequest indexRequest = new IndexRequest();
+		indexRequest.source("{}", XContentType.JSON);
+		UpdateQuery updateQuery = new UpdateQueryBuilder().withId(randomNumeric(5))
+				.withClass(SampleEntity.class).withIndexRequest(indexRequest).build();
+		elasticsearchTemplate.update(updateQuery);
+	}
+}
